#24 Created testmaps folder for testing, and started working on the TournamentSelection unit tests. Modified the randomstub to generate doubles instead of floats, and changed the tournamentselection class to use doubles when rolling for probability

+++ a1-tsp/src/main/java/uni/evocomp/a1/TournamentSelection.java
@@ -69,7 +69,7 @@ public class TournamentSelection implements SelectSurvivors {
     // Choose the best individual with probability p, 2nd with p(1-p), 3rd with p(1-p)^2 etc.
     int i = 0;
     for (Individual individual : tournamentList) {
-      if (rand.nextFloat() < (p * Math.pow((1 - p), i))) {
+      if (rand.nextDouble() < (p * Math.pow((1 - p), i))) {
         survivors.add(individual);
         if (survivors.size() < (int) (survivalProportion * population.size)) {
           break;
